:sparkles: Changes to API
